Simplify startup instructions
Instead of requiring a separate 'run.sh' (which only takes care of
launching the container, not building it), let's have a
docker-compose.yml configuration which builds the image if needed *and*
comes with a default command.

So after checking out (and having Docker installer), you can just do

  docker compose run --rm app

And that's it!