Turn Server state into a real map
I plan to extend the state managed by a server so that it's not just the
Session but also a timer keeping track of the round time. Let's prepare
that step by first turning the state into a real map. Maybe I should
have done that right from the start...